dom/media/test/test_background_video_no_suspend_short_vid.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>Test Background Video Doesn't Suspend When Timeout Is Longer Than Video</title>
<script src="/tests/SimpleTest/SimpleTest.js"></script>
<script src="manifest.js"></script>
<script src="background_video.js"></script>
<link rel="stylesheet" href="/tests/SimpleTest/test.css"/>
<script>
"use strict";

var manager = new MediaTestManager;

startTest({
  desc: "Test Background Video Doesn't Suspend When Timeout Is Longer Than Video.",
  prefs: [
    [ 'media.test.video-suspend', true ],
    [ 'media.suspend-bkgnd-video.enabled', true ],
    // Gizmo.mp4 is about 5.6s
    [ 'media.suspend-bkgnd-video.delay-ms', 10000 ]
  ],
  tests: gDecodeSuspendTests,
  runTest: (test, token) => {
    let v = appendVideoToDoc(test.name, token);
    manager.started(token);

    /* This test checks that suspend doesn't occur when the delay is longer
       than the duration of the video that's playing */
    waitUntilPlaying(v)
      .then(() => checkVideoDoesntSuspend(v))
      .then(() => {
        ok(true, 'Video ended before decode was suspended');
        manager.finished(token); })
      .catch((e) => {
        ok(false, 'Test Failed: ' + e.toString());
        manager.finished(token); });
  }
});
</script>
